+++ gobblin-restli/gobblin-flow-config-service/gobblin-flow-config-service-server/src/test/java/org/apache/gobblin/service/FlowExecutionResourceHandlerTest.java
@@ -21,7 +21,7 @@ import org.testng.Assert;
 import org.testng.annotations.Test;
 
 
-public class FlowExecutionResourceLocalHandlerTest {
+public class FlowExecutionResourceHandlerTest {
 
   @Test
   public void testEstimateCopyTimeLeftSanityCheck() {
